Run the compileTestTask sooner
Currently the plugin executes all its tasks before the check task. This
is a problem when the user wants to run a test from the IDE, because the
test classes are not compiled.

Indeed, the IDE runs only the assemble* task before launching a junit,
so I simply moved the compileTestTask after the corresponding compile*
task. As a result the test classes are well compiled before the junit is
executed.

Note: I used `Task.finalizedBy()` which is only available since Gradle 1.7,
so I have to upgrade the wrapper version.